Split inventory scanning by asset type
